Apply the warning policy to every target, and fix what it found
The policy was a foreach over four hand-written target names sitting below
the product declarations. Seven products and the update helper were added
above it and none of them inherited it, so -Wno-nontrivial-memcall reached
1 of the 9 targets that compile JUCE's vendored harfbuzz and the other 8
each emitted ~22 warnings about third-party code on every clean build.

It now lives in px3_add_product, so a product added tomorrow cannot opt out
of it by being added, and in an explicit list for the console targets that
function does not build.

Turning it on everywhere surfaced real things:

- A dangling `for (auto* child : editor->getChildren())` whose body was the
  next statement, calling refreshFromParameters() once per child.
- AmpEnvDsp_IgnoresHoldEntirely compared a trace with itself. Its helper
  took a hold argument and never applied it - EnvelopeSettings has no such
  field - so both renders came from identical settings. The absence of a
  hold stage is now checked against the type, which fails the moment
  EnvelopeSettings grows one.
- A vibe survey in TestsMain was nested inside the noise-floor block by
  accident, shadowing its processor and capture.

build-product.sh gets two fixes while its neighbours are open. Asking for
no format built the bare target, which juce_add_plugin makes the shared-code
static library: it compiled everything, linked the .a, and produced a bundle
skeleton with no executable that macOS refuses to open. It now builds
${TARGET}_All. And --debug meant CMAKE_BUILD_TYPE=Debug here while meaning
the in-plugin debug panel in build-release.sh; it now means the panel in
both, with --config Debug for an unoptimised build.

run-standalone.sh's staleness check globbed Source/, which the ecosystem
migration moved. find failed into 2>/dev/null, the timestamp came back
empty, and the check passed on every stale build it existed to catch.

+++ updater/UpdaterMain.cpp
@@ -134,7 +134,9 @@ int runInstaller(const juce::File& installer)
     }
 
     process.waitForProcessToFinish(-1);
-    return process.getExitCode();
+    // getExitCode returns uint32; a process exit status is a byte, so the
+    // narrowing is safe - it is made explicit rather than implicit.
+    return static_cast<int>(process.getExitCode());
 }
 } // namespace
